+++ R/topNList.R
@@ -36,6 +36,24 @@ setMethod("getList", signature(from = "topNList"),
 
 setAs("topNList", "list", function(from) getList(from, decode = TRUE))
 
+setMethod("c", signature(x = "topNList"),
+  function(x, ..., recursive = FALSE){
+    args <- list(...)
+    if (recursive)
+      args <- unlist(args)
+    for (y in args) {
+      if (!is(y, "topNList"))
+        stop("can only combine objects of class topNList")
+      if (!identical(x@itemLabels, y@itemLabels))
+        stop("topNlists are not compatible (item labels do not match).")
+      x@items <- c(x@items, y@items)
+      x@ratings <- c(x@ratings, y@ratings)
+    }
+    x
+  }
+)
+
+
 ## creation from realRatingMatrix
 setMethod("getTopNLists", signature(x = "realRatingMatrix"),
   function(x, n=10, randomize=NULL, minRating=NA){